[nrf noup] wifi_mgmt: Add support for non-offload TCP/IP
Wi-Fi offload has two parts control and data, when wifi_mgmt is used
for control the data part if assumed to be offloaded (WIFI/NET_OFFLOAD),
but some drivers do not support that assumption.

Extend the wifi_mgmt to take in a full Ethernet APIs so that wifi_mgmt
can be used with native TCP/IP stack.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>