[nrf fromtree] testsuite: ztress: First sys_rand_get32 from thread
Added a workaround to call random generator once during the
initialization. It is done to handle XOSHIRO generator limitation
which performs initialization in the first sys_rand32_get call.
And for some entropy generators it cannot be done from an interrupt
context and it happens if k_timer context is used which expires first.

Cherry-picked from f3ea32b01d.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>