[nrf noup] net: ip: Fix nRF Wi-Fi management events
When using nRF Wi-Fi management it uses the net management events with
information, but the size of the events is not enough to hold the Wi-Fi
events as the proper size is defined only for Zephyr's Wi-Fi management.

Use the proper size for both nRF and Zephyr's Wi-Fi management.

Note: If CONFIG_NET_DHCPV4 is enabled, then this works, as size is now large
enough for Wi-Fi management events.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>
(cherry picked from commit 98bc3a7a6b3917ce5a058e7a7271d3036edffb88)
Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>