[nrf noup] cmake: add ncs_unity CMake module
This commit adds a nRF Connect SDK unity CMake module which allows the
use of unity and CMake together with the unit_testing board.

This file is created as a `noup` patch, which allows it to be easily
identified and reverted when Zephyr enhanced `unittest` CMake module
to properly process out-of-tree extensions.

When Zephyr provides better use of unittest, then the file
`ncs_unity.cmake` shall be moved to sdk-nrf repository.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>