[nrf fromlist] modules: mbedtls: Provide `mbedtls_hardware_poll()` conditionally
Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/51362.

This is a follow-up to commit c7327f5f707bfb367a78be05325469f2e950171a.

Wrap implementation of the `mbedtls_hardware_poll()` function in
`#if defined(CONFIG_MBEDTLS_ZEPHYR_ENTROPY)` so that the function
is provided only when that option is activated.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>