[nrf noup] samples: psa_crypto: update documentation
The sample is requesting an initial attestation token, which requires
running the provisioning sample before, so the device has the needed
key material programmed.
This updates the documentation to make users aware of this and as the
provisioning sample is in sdk-nrf, a noup is used.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
(cherry picked from commit 9114a18539af6eb880a528b49e1e5e11b8b8af32)