[nrf noup] samples&tests: Restore a few CONFIG_NEWLIB_LIBC_NANO=n
This is a follow-up to commit 9dd570f8a2207a02c886480b0065cd5a0e3a2e94.

Since in NCS, unlike in vanilla Zephyr, the nano variant of newlib
is the default one, restore entries that disable the nano variant
in one sample and one test that require the full newlib variant.
This patch is supposed to be removed when picolibc becomes the default.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>