[nrf fromtree] nrf52_bsim: Add option to delay CPU and test initialization
In some simulated test cases, it can be usefull
to delay the boot of the CPU and the test initialization
for a while. Namely by the offset of the device
relative to other simulated devices.

We add the option to select so from command line.
By default the option is disabled so this change is
backwards compatible.

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>
(cherry picked from commit fb95defce3a944e2bdfb0d1d5ce20585195c88a5)