[nrf noup] boards: thingy53_nrf5340: Set MCUBoot to use all of RAM
Change enables MCUBoot to use all available RAM on thingy53 as it require
more RAM than what is reserved as secure. This will make thingy53 a little
bit less secure than our other boards. Since this is a prototype board I
think it should be okay as we require the enabled MCUBoot features for it
to be programmable without a debugger.

Jira: NCSDK-17250

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>