[nrf noup] scripts: west_commands: runners: nrfjprog: bring --snr back
This patch partially reverts 2cee5ff.

nRF VScode extension implemented support for the new option, `--dev-id`,
however, it relied on deprecated code paths to detect wether Zephyr
supported `--dev-id` or not. The extension has been updated to fix this
issue, so this patch can be safely removed once a new release of the
extension gets deployed. In the mean time, it will allow people working
on `main` to not experience any issues.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>