modules: canopennode: Updated CANopenNode driver to v2.0
This commit updates the existing zephyr driver for canopennode from v1.3 to
v2.0 and adds support for all new upstream driver features.
The motivation for this was that I needed the ASCII gateway feature, which
is not available in v1.3.
I ran checkpatch on my own code, but the stuff in the objdict-directory of
the example is autogenerated and won't pass. But I hardly even touched that,
so it should be fine.

Signed-off-by: Alexander Tuxen <tuxen@silpion.de>