[nrf fromtree] doc: Log unused log filter patterns
This allows us to identify any patterns that we can remove or that
we didn't realise are no longer in use. This might happen if issues
within doxygen or docleaf are resolved.

This allows us to remove the pattern:

  .*Duplicate C declaration.*\n.*'\.\. c:.*:: uint16_t id'.*

which does not match anything in the current set up.

We also split the filter patterns in known-warnings out into different
sections depending on their cause.

Also extend the pattern parser to ignore empty lines so that we can
have some formatting in the known-warnings file.

fromtree notes: slightly adjusted, docleaf changes not imported

Signed-off-by: Michael Jones <m.pricejones@gmail.com>
Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
(cherry picked from commit 34e50037f0d94efce10770da5abcf48f7ba33906)