[nrf fromlist] drivers: spi: Fix crash for large buffers
In case the chunk size is larger than the configured RAM size then for
haltium we end up writing beyond the buffer causing a bus fault.

Fix the checks for chunks to include even if the buffer is in RAM as
bounce buffer is always used in nRF54H series SoC's.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/69081

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>