[nrf fromtree] twister: Fix error for --device-testing with not runnable integration
An urforseen error was introduce with #62713. When "--device-testing"
or "--filter runnable" is used in CLI, twister will skip tests which
cannot be executed (not just built) on a given platform. If a given
platform is among integration platforms it will cause an error.
However, it shouldn't be the case. Such skip is intentional and
shouldn't be an error. Fixed by assigning an existing Filter.CMD_LINE,
which is exempted from being turned to error.

fixes: #63844

(cherry picked from commit 4cd7df2eec31969e84cae74cb49b0bb5710c200d)
Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>