[nrf fromlist] samples: soc_flash_nrf: Make sure that the flash partition is valid
Make sure that the flash partition is valid.
With TF-M enabled the storage_partition can be used as memory that is
known to be configured as non-secure flash region.
The slot1_ns_partition partition is only correct when TF-M is built
with BL2 enabled.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/66279

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>