[nrf fromtree] net: tcp: Install a last ack timer in passive close
If we are in a passive close state, then it is possible that
the ack we are waiting is lost or we do not accept the one peer
sent to us because of some earlier out of memory issue.
So install a timer (using by default the FIN timer value) to
close the connection if the last ack is not received on time.

Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
(cherry picked from commit 40215e07a39305125d47f04f6b4a498b71d79c43)
Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>