[nrf fromlist] wifi: shell: Fix the channel extraction
The channel extraction from string directly uses the end variable with
limited data type, this causes issue if an invalid channel that exceeds
the data is given as an input e.g., 300, which would end up as a valid
channel 44.

Use an intermediate variable with type that can hold all possible
combinations (valid and invalid) and only after validation assign that
to the end type.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/67694

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>