[nrf fromlist] samples: Bluetooth: Only connect to devices with rssi below -50
When the Bluetooth central samples in an open air environment
it is very likely that there are multiple devices nearby with
a received signal strength stronger than -70 dBm.

To avoid connecting to the wrong device, make the check stricter.

For tests using those samples, the NtNcable attenuation is changed
from the default 60 dBm to 40 dBm to satisfy the new requirement.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/68870

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>